service/requirements.txt
aiohttp
pep8==1.7.0
pymongo==2.8.0
pytest==2.8.5
motor==0.5b0
aiohttp-jinja2
yaml-config==0.1.1